07_rag_knowledge_base/requirements.txt
langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.3
llama-index==0.10.26
pinecone-client==3.2.2
faiss-cpu==1.8.0
sentence-transformers==2.7.0
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.33.0
openai==1.23.0
tiktoken==0.6.0
pypdf==4.1.0
python-dotenv==1.0.1
